Future lock package merged into lock (#16565)
