+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/resources/META-INF/services/org.apache.shardingsphere.mode.manager.cluster.coordinator.registry.GovernanceWatcher
@@ -21,5 +21,5 @@ org.apache.shardingsphere.mode.manager.cluster.coordinator.registry.config.watch
 org.apache.shardingsphere.mode.manager.cluster.coordinator.registry.config.watcher.PropertiesChangedWatcher
 org.apache.shardingsphere.mode.manager.cluster.coordinator.registry.status.compute.watcher.ComputeNodeStateChangedWatcher
 org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.watcher.LockChangedWatcher
-org.apache.shardingsphere.mode.manager.cluster.coordinator.future.lock.watcher.GlobalLocksChangedWatcher
-org.apache.shardingsphere.mode.manager.cluster.coordinator.future.lock.watcher.GlobalAckChangedWatcher
+org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.watcher.GlobalLocksChangedWatcher
+org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.watcher.GlobalAckChangedWatcher
